requirements.txt
# Core
soundcard==0.4.5
soundfile==0.13.1
numpy==1.26.4

# NotebookLM API
notebooklm-py[browser]==0.3.3

# Auth (login via Chrome CDP)
playwright==1.58.0

# Build
pyinstaller==6.19.0
